+++ crates/swc/src/config/mod.rs
@@ -466,7 +466,7 @@ impl Options {
                 swc_plugin_runner::cache::init_plugin_module_cache_once(&experimental.cache_root);
             }
 
-            crate::plugin::plugins(plugin_resolver, experimental, plugin_context)
+            crate::plugin::plugins(Some(plugin_resolver), experimental, plugin_context)
         };
 
         // Native runtime plugin target, based on assumption we have
@@ -486,7 +486,8 @@ impl Options {
                 env_name: self.env_name.to_owned(),
             };
 
-            crate::plugin::plugins(experimental, plugin_context)
+            swc_plugin_runner::cache::init_plugin_module_cache_once();
+            crate::plugin::plugins(None, experimental, plugin_context)
         };
 
         #[cfg(not(feature = "plugin"))]