+++ crates/swc_css_prefixer/src/lib.rs
@@ -1,0 +1,3 @@
+pub use self::prefixer::prefixer;
+
+mod prefixer;